Changes that at the end of the day may not have been needed:
single item vs. itemList overloads were not working (single item was being selected for ICollection types), but this may have been due to the fact that the call was not using generic method signature.
"propertyInfo" was replaced with "memberInfo", but that too was not neccessary most likely, because the actual bug was caused by bad string comparison
I am committing, but consider rolling back